+++ sendvalues.py
@@ -4,6 +4,7 @@ import datetime
 import os
 import requests
 import time
+import statistics
 
 bus = smbus2.SMBus(1)
 lastTemp = -300.0
@@ -13,18 +14,21 @@ lastHumUpdate = datetime.datetime(2024, 1, 1, 0, 0)
 
 while True:
     loopStart = datetime.datetime.utcnow()
-    tempSum = 0.0
-    humSum = 0.0
+    tempSamples = []
+    humSamples = []
     
-    for x in range(0, 5):
+    for x in range(0, 10):
         bus.write_i2c_block_data(0x44, 0x2C, [0x06])
         data = bus.read_i2c_block_data(0x44, 0x00, 6)
-        tempSum += ((((data[0] * 256.0) + data[1]) * 175) / 65535.0) - 45
-        humSum += 100 * (data[3] * 256 + data[4]) / 65535.0
+        sampleTemp = ((((data[0] * 256.0) + data[1]) * 175) / 65535.0) - 45
+        tempSamples.append(sampleTemp)
+        sampleHum = 100 * (data[3] * 256 + data[4]) / 65535.0
+        humSamples.append(sampleHum)
         time.sleep(1)
     
-    cTemp = round(tempSum / 5.0, 2)
-    humidity = round(humSum / 5.0, 1)
+    cTemp = round(statistics.harmonic_mean(tempSamples), 2)
+    humidity = round(statistics.harmonic_mean(humSamples), 1)
+
     now = datetime.datetime.utcnow()
     timestamp = now.isoformat()
     token = os.environ.get('SUPERVISOR_TOKEN','?')
@@ -32,20 +36,30 @@ while True:
     
     tempTimeDelta = (now - lastTempUpdate)
     if lastTemp != cTemp or tempTimeDelta.total_seconds() > 300:
-        url = 'http://supervisor/core/api/states/sensor.sht30_temperature'
-        data = {'entity': 'sensor.sht30_temperature', 'attributes': { 'friendly_name': 'SHT30 Temperature', 'unit_of_measurement': '\N{DEGREE SIGN}C'}}
+        entityId = 'sensor.sht30_temperature'
+        url = 'http://supervisor/core/api/states/' + entityId
+        data = {'attributes': { 'friendly_name': 'SHT30 Temperature', 'unit_of_measurement': '\N{DEGREE SIGN}C'}}
+        data['entity'] = entityId
         data['state'] = cTemp
         data['last_updated'] = timestamp
+        data['attributes']['stdev'] = round(statistics.stdev(tempSamples), 3)
+        if lastTemp > -274:
+            data['attributes']['delta'] = cTemp - lastTemp
         response = requests.post(url, data=json.dumps(data), headers=headers)
         lastTemp = cTemp
         lastTempUpdate = now
     
     humTimeDelta = (now - lastHumUpdate)
     if lastHum != humidity or humTimeDelta.total_seconds() > 300:
-        url = 'http://supervisor/core/api/states/sensor.sht30_humidity'
-        data = {'entity': 'sensor.sht30_humidity', 'attributes': { 'friendly_name': 'SHT30 Relative Humidity', 'unit_of_measurement': '%'}}
+        entityId = 'sensor.sht30_humidity'
+        url = 'http://supervisor/core/api/states/' + entityId
+        data = {'attributes': { 'friendly_name': 'SHT30 Relative Humidity', 'unit_of_measurement': '%'}}
+        data['entity'] = entityId
         data['state'] = humidity
         data['last_updated'] = timestamp
+        data['attributes']['stdev'] = round(statistics.stdev(humSamples), 3)
+        if lastHum >= 0 and lastHum <= 100:
+	    data['attributes']['delta'] = humidity - lastHum
         response = requests.post(url, data=json.dumps(data), headers=headers)
         lastHum = humidity
         lastHumUpdate = now